Add UI and Images
Add images for rock paper and scissors to be displayed on cards for the
player and the computer. Clickable buttons reveal each selection and
update the game score. All the scoring is only shown in the console, and
the game ends (and click listeners on buttons removed) once one player
reaches 5 wins.